+++ CHANGELOG.md
@@ -17,6 +17,7 @@ _Most_ notable changes to this project will be documented in this file.
 - Cardboard Box 1 block.
 - Microscope block.
 - Office Desk block.
+- Panel Fan block.
 - Printer block.
 - Radio block.
 - Sink block.